HIVE-29554: Refactor MTableColumnStatistics and MPartitionColumnStatistics to avoid code duplication
